+++ server/internal/logic/admin/post.go
@@ -153,7 +153,7 @@ func (s *sAdminPost) View(ctx context.Context, in adminin.PostViewInp) (res *adm
 }
 
 // List 获取列表
-func (s *sAdminPost) List(ctx context.Context, in adminin.PostListInp) (list []*adminin.PostListModel, totalCount int, err error) {
+func (s *sAdminPost) List(ctx context.Context, in adminin.PostListInp) (list []*adminin.PostListModel, totalCount int64, err error) {
 	mod := dao.AdminPost.Ctx(ctx)
 
 	// 访问路径
@@ -181,7 +181,7 @@ func (s *sAdminPost) List(ctx context.Context, in adminin.PostListInp) (list []*
 		return list, totalCount, nil
 	}
 
-	if err = mod.Page(in.Page, in.PerPage).Order("id desc").Scan(&list); err != nil {
+	if err = mod.Page(int(in.Page), int(in.PerPage)).Order("id desc").Scan(&list); err != nil {
 		err = gerror.Wrap(err, consts.ErrorORM)
 		return list, totalCount, err
 	}
